6871: Moved button code to separate files
