+++ assets/admin/translations/da/common.json
@@ -1249,6 +1249,6 @@
   "screen-columns": {
     "screen-groups-modal-title": "Skærmgrupper",
     "campaigns-modal-title": "Kampagner",
-    "active": "aktive"
+    "active": "Aktiv"
   }
 }
